Move admin ListTable abstract.
